+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 * 2019-03-07
   * Added support for fetching MT940 reports (Marcin Orlowski)
-  
+  * Now throws \RuntimeException in place of too generic \Exception (Marcin Orlowski)